feat : 코드 리뷰 반영 수정
Button의 interface 수정, AdmissionHeader, CommonHeader을 BaseHeader로 합치기
