+++ libs/ui/src/lib/index.ts
@@ -1,4 +1,3 @@
-export * from './CommonHeader';
 export * from './icons';
-export * from './AdmissionHeader';
 export * from './Button';
+export * from './BaseHeader';
